+++ src/composables/service-helpers.ts
@@ -1,9 +1,9 @@
 import { getDisplayedOsidbError } from '@/services/OsidbAuthService';
 import { useToastStore } from '@/stores/ToastStore';
-import { AxiosError, type AxiosResponse } from 'axios';
+
 
 export function createCatchHandler(title: string = 'Error', callback?: () => void, shouldThrow: boolean = true) {
-  return (error: AxiosError) => {
+  return (error: any) => {
     const displayedError = getDisplayedOsidbError(error);
     const { addToast } = useToastStore();
     addToast({
@@ -27,7 +27,7 @@ export function createSuccessHandler({
   title = 'Operation Successful',
   body,
 }: Record<string, string>) {
-  return (response: AxiosResponse | void) => {
+  return (response: {data: any}) => {
     const { addToast } = useToastStore();
     addToast({
       title,